+++ algoritmo_knn.h
@@ -1,13 +1,13 @@
 typedef struct registro {
-    int gender; // 0 = Hombre, 1 = Mujer
+    float gender; // 0 = Hombre, 1 = Mujer
     float age;
-    int hypertension;
-    int heart_disease;
-    int smoking_history;
+    float hypertension;
+    float heart_disease;
+    float smoking_history;
     float bmi;
     float HbA1c_level;
-    int blood_glucose_level;
-    int diabetes; // 0 = No tiene, 1 = Tiene
+    float blood_glucose_level;
+    float diabetes; // 0 = No tiene, 1 = Tiene
 } Registro;
 
 void normalizar_registro(Registro*);
